Added styling and logic to different pages
Many pages needed some styling. The research bars needed some styling for smaller screens and I needed to implement logic to show only events that are happening either today or in the future. I tried to counters and filters in the html, but it's been impossible, so I started searching for documentations to add functions to the models and that's what I ended up doing